fix seedless show recovery methods
