+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/accountSettings/seedlessExportPhrase/readyToExport.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback, useEffect } from 'react'
 import { useRouter } from 'expo-router'
 import { useFocusEffect, useNavigation } from '@react-navigation/native'
-import { showAlert, View } from '@avalabs/k2-alpine'
+import { Icons, showAlert, useTheme, View } from '@avalabs/k2-alpine'
 import Logger from 'utils/Logger'
 import AnalyticsService from 'services/analytics/AnalyticsService'
 import BackBarButton from 'common/components/BackBarButton'
@@ -12,6 +12,7 @@ import {
 } from 'features/accountSettings/context/SeedlessMnemonicExportProvider'
 import { Platform } from 'react-native'
 import { SeedlessExportMnemonicPhrase } from 'features/accountSettings/components/SeedlessExportMnemonicPhrase'
+import NavigationBarButton from 'common/components/NavigationBarButton'
 
 const SeedlessExportReadyScreen = (): JSX.Element => {
   const { back, canGoBack } = useRouter()
@@ -19,6 +20,9 @@ const SeedlessExportReadyScreen = (): JSX.Element => {
     useSeedlessMnemonicExportContext()
   const { getParent } = useNavigation()
   const [hideMnemonic, setHideMnemonic] = useState(true)
+  const {
+    theme: { colors }
+  } = useTheme()
 
   const onCancelExportRequest = useCallback(
     ({ title, description }: { title: string; description: string }): void => {
@@ -84,10 +88,36 @@ const SeedlessExportReadyScreen = (): JSX.Element => {
     [customGoBack]
   )
 
+  const renderHeaderRight = useCallback(() => {
+    return (
+      <NavigationBarButton>
+        {hideMnemonic ? (
+          <Icons.Action.VisibilityOff
+            color={colors.$textSecondary}
+            onPress={toggleRecoveryPhrase}
+            hitSlop={16}
+          />
+        ) : (
+          <Icons.Action.VisibilityOn
+            color={colors.$textPrimary}
+            onPress={toggleRecoveryPhrase}
+            hitSlop={16}
+          />
+        )}
+      </NavigationBarButton>
+    )
+  }, [
+    colors.$textPrimary,
+    colors.$textSecondary,
+    hideMnemonic,
+    toggleRecoveryPhrase
+  ])
+
   useFocusEffect(
     useCallback(() => {
       getParent()?.setOptions({
-        headerLeft: renderCustomBackButton
+        headerLeft: renderCustomBackButton,
+        headerRight: renderHeaderRight
       })
 
       return () => {
@@ -95,13 +125,12 @@ const SeedlessExportReadyScreen = (): JSX.Element => {
           headerLeft: undefined
         })
       }
-    }, [getParent, renderCustomBackButton])
+    }, [getParent, renderCustomBackButton, renderHeaderRight])
   )
 
   return (
     <SeedlessExportMnemonicPhrase
       mnemonic={mnemonic}
-      toggleRecoveryPhrase={toggleRecoveryPhrase}
       hideMnemonic={hideMnemonic || mnemonic === undefined}
       onCopyPhrase={handleCopyPhrase}
     />
